* FvwmButtons/draw.c (RedrawButton): fixed background pixmap for swallowed windows. Is it ok to never draw a background for swallowed applications?
* FvwmIconMan/fvwm.c (destroy_window): coredump fixed: IconMan died
when the window with the focus was deleted and the focus was changed

* FvwmButtons/parse.c (ParseContainer):
(match_string): fixed possible coredump (font_string == NULL but
b_Font was still set).

* fvwm/read.c (ReadSubFunc): the piperead string was set as the last
read filename

* fvwm/module.c (executeModule): removed useless code:
  if(args[nargs] != NULL)
    free(args[nargs]);
when args[nargs] is guaranteed to be NULL

* fvwm/fvwm.c (SetRCDefaults): added missing comma after "SetMenuStyle"
default

* fvwm/builtins.c (SetMenuStyle):
* fvwm/fvwm2.1: removed SetMenuStyle ... next. The menuface option can
be applied to any menu style. I don't understand what this was good for
anyway since when you specified 'next', the menu style itself
(mwm/fvwm/win) was undefined.

* fvwm/builtins.c (SetMenuStyle1): removed unused code