Many bugfixes and code cleanup.
The biggest fix was to restore the ability to read
FVWM_CONFIGDIR/system.fvwm2rc file, which got lost sometime in the
past week or two.

Rewrote CatString3 to dynamically allocate the buffer, so it is more
generally useful.  If you find yourself doing lots of mallocs &
strcats, check this function out!