Try to fix "variablility" in size of windows in pager; mini-window width and height now calculated by scaling original windows' width and height. Also, since calculations are performed using integer arithmetic, do all multiplications before the final division to prevent division inaccuracies from being scaled up.
NOTE: this has reduced but not removed the "variablility" I am
seeing. The pager icon is still sometimes wrong when I minimise an
application and the animation has completed.

e.g.: the mini-window's size should be 1x1, which is converted to 3x3
to make it visible. However, the animated mini-window iconifies to
4x1, which is converted to 4x3. (OTOH, it used to 5x3 occasionally.)