Correctly restore windows coming out of fullscreen
When windows enter into fullscreen, ensure certain attributes of that
window -- such as the iconification, maximisation, shading, etc., are
tracked and correctly restored when the window leaves fullscreen mode.

In the case of window maximisation, because fullscreen mode internally
uses "ResizeMoveMaximize" to achieve this, the effect of this is to lose
the window's maximised state in the interim, because calling
ResizeMoveMaximize will remove the maximised state of the window before
changing it for its own purposes.