Add TitleFormat and IconTitleFormat style options
This allows for free-form visible names for both iconic and non-iconic
windows, with various placeholders to fill in information such as the
window's name or class.

In doing so, depecates:

IndexedWindowName, IndexedIconName, ExactWindowName,
ExactIconName

This might help people out using clients such as Firefox, to be able to more
uniquely distinguish between one window and the other, where the WM_CLASS is
often the same, etc.