Enables desks/pages to be per-screen
This augments the struct monitor instance to hold information about
pages/desks to be per-monitor.

Because of the way paging works, it's recommended that desks are used, and
setting "DesktopSize 1x1" doe now, so that no paging occurs, only Desk
switching (which internally unmaps the windows, rather than move them, which
is what paging does.)

FvwmPager is currently broken as well; it doesn't have a concept of
per-monitor windows, either.