Paging:  Make callers monitor-context aware
When switching pages/desks, don't assume the desk/page in question is the
monitor which has the mouse pointer.  It's entirely possible the context for
the monitor are from the window (as in a ClientMessage, for instance).

Therfore, make all places where switching pages/desks happens to pass in a
struct monitor *, and use that.