FScreenGetScrRect: Remove m == NULL check (GCC fix)
Because FScreenGetScrRect() initialises values passed in for coordinates,
don't test to see if the given monitor is NULL (which would have
resulted in the premature exit of this function, potentially leaving the
values unassigned).

We always guarantee finding a given monitor, and hence can set these values
correctly.  The original check was bogus, and caused GCC to rightly complain
about potential uninitialised values (-Werror=maybe-uninitialized).