Redefine how the "global" monitor is stored
Rather than making the "global" monitor part of the monitor queue in the case
where XRandR is in use, instead only add the global monitor to the queue if
XRandR isn't in use (and hence there's only ever one monitor).

This means the callers don't need to worry, when querying monitors, about the
global monitor since contextually, querying monitors is always going to be
correct.  If the global monitor is required, that can be queried in the same way
it is now, and hence those existing callers don't need to change.

In making this change, initialise the monitors when required, rather than based
on whether they're in the main monitor_q or not; meaning no special-casing of
the global monitor is needed to support initialisation.